@@ -1,3 +1,3 @@
 module github.com/lakal3/dllcall
 
-go 1.12
+go 1.21
