@@ -4,5 +4,7 @@ package main
 #cmethod Greet
 */
 type greeting struct {
-	greeting string
+	text string
+	data []byte
+	next *greeting
 }
