@@ -9,6 +9,7 @@ import (
 
 var fKeep bool
 var fFastcall bool
+var fPin bool
 
 func main() {
 	err := sys_init()
@@ -17,6 +18,7 @@ func main() {
 	}
 	flag.BoolVar(&fKeep, "keep", false, "Keep temp generator file")
 	flag.BoolVar(&fFastcall, "fast", false, "Use fastcall for csafe_method(s)")
+	flag.BoolVar(&fPin, "pin", false, "Pin memory for inner pointers. This feature required go1.21")
 	flag.Parse()
 	if flag.NArg() < 2 {
 		usage()
@@ -35,7 +37,7 @@ func main() {
 }
 
 func usage() {
-	fmt.Println("dllcall v0.8.2")
+	fmt.Println("dllcall v0.9.1")
 	fmt.Println("Usage: dllcall [flags] go_file cpp_file")
 	os.Exit(1)
 }
